backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.0
requests==2.31.0
python-decouple==3.8
numpy==1.26.3
scipy==1.11.4